chore: renames style files
